functions/start-instance/requirements.txt
google-cloud-storage==1.29.0
google-compute-engine==2.8.13
